+++ spec/jkf/converter/csa_spec.rb
@@ -53,7 +53,7 @@ describe Jkf::Converter::Csa do
 
     context 'when 8mai' do
       let(:kif_parser) { Jkf::Parser::Kif.new }
-      let(:filename) { fixtures(:kif).detect { |file| file =~ /8mai/ } }
+      let(:filename) { fixtures(:kif).detect { |file| file.include?('8mai') } }
       let(:jkf) { kif_parser.parse(File.read(filename, encoding: 'Shift_JIS').toutf8) }
 
       it { expect(csa_converter.convert(jkf)).to be_a String }