+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Others
 
+* Lint with RuboCop performance plugin
 * Update documentation
 
 ## [0.5.1] - 2023-06-24